UIDEXP-29: Fix according to the code review (without tests)
