+++ src/components/JobLogsContainer/fetchFileDownloadLink.js
@@ -1,6 +1,6 @@
 import { createOkapiHeaders } from '@folio/stripes-data-transfer-components';
 
-export const getFileDownloadLinkRequest = async (jobLog, okapi) => {
+export default async (jobLog, okapi) => {
   const { url: host } = okapi;
   const url = `${host}/data-export/jobExecutions/${jobLog.id}/download/${jobLog.exportedFiles[0].fileId}`;
   const response = await fetch(url, {
