+++ src/utils/index.js
@@ -1,0 +1,2 @@
+export * from './downloadFileByLink';
+export * from './constants';
