+++ test/bigtest/tests/JobLogs-test.js
@@ -11,7 +11,7 @@ import { jobLogsContainerInteractor } from '../interactors';
 
 const getCellContent = (row, cell) => jobLogsContainerInteractor.logsList.rows(row).cells(cell).content;
 
-describe('Job logs list', () => {
+describe.only('Job logs list', () => {
   describe('default scenario', () => {
     setupApplication();
 
@@ -59,34 +59,6 @@ describe('Job logs list', () => {
         it('should have the correct query in path', function () {
           expect(this.location.search.includes('direction=descending&sort=status')).to.be.true;
         });
-      });
-    });
-
-    describe('clicking on file name with success flow', () => {
-      beforeEach(async () => {
-        await jobLogsContainerInteractor.fileNameBtns(0).click();
-      });
-
-      it('should not show error notification', () => {
-        expect(jobLogsContainerInteractor.callout.errorCalloutIsPresent).to.be.false;
-      });
-    });
-  });
-
-  describe('error scenario', () => {
-    setupApplication({ scenarios: ['fetch-get-download-link-error'] });
-
-    beforeEach(function () {
-      this.visit('/data-export');
-    });
-
-    describe('clicking on file name with error flow', () => {
-      beforeEach(async () => {
-        await jobLogsContainerInteractor.fileNameBtns(0).click();
-      });
-
-      it('should show error notification', () => {
-        expect(jobLogsContainerInteractor.callout.errorCalloutIsPresent).to.be.true;
       });
     });
   });
